+++ src/app/constants/number.constants.ts
@@ -3,5 +3,7 @@ export const MAGIC_NUMBERS = {
     N_0: 0,
     N_1: 1,
     N_2: 2,
-    N_3: 3
+    N_3: 3,
+    N_6: 6,
+    N_9: 9
 };