+++ src/app/controller/characterController.ts
@@ -1,3 +1,4 @@
+import { FormControl } from '@angular/forms';
 import { Character } from 'src/app/model/character';
 import { MAGIC_NUMBERS } from '../constants/number.constants';
 import { Atributes } from '../model/atributes';
@@ -9,6 +10,7 @@ import { GodAffinity } from '../model/godAffinity';
 import { Modificator } from '../model/modificator';
 import { Power } from '../model/power';
 import { Skill } from '../model/skill';
+import { NAMES } from './character.constants';
 
 export class CharacterController{
 
@@ -107,6 +109,11 @@ export class CharacterController{
       newCharacter.setHistory(protoCharacter.history);
 
       return newCharacter;
+  }
+
+  public static getRandomName(): string {
+    const randomIndex = Math.floor(Math.random() * NAMES.length);
+    return NAMES[randomIndex];
   }
 
   private static copyCombatEquip(combatEq: any[]): CombatEquip[]{